kenan:chore(webapp)=implementation ajout Etape + modif + supp

+++ Backoffice/rando/src/main/java/com/rando/dto/EtapeDto.java
@@ -10,9 +10,9 @@ public class EtapeDto {
 	@NotBlank(message = "Vous devez fournir une description !")
 	private String description;
 	@NotBlank(message = "Vous devez indiquer une latitude !")
-	private BigDecimal latitude;
+	private Integer latitude;
 	@NotBlank(message = "Vous devez indiquer une longitude !")
-	private BigDecimal longitude;
+	private Integer longitude;
 	@NotBlank(message = "Vous devez donner un nom à cette étape !")
 	private String nom;
 	private String qrCode;
@@ -45,25 +45,25 @@ public class EtapeDto {
 	/**
 	 * @return the latitude
 	 */
-	public BigDecimal getLatitude() {
+	public Integer getLatitude() {
 		return latitude;
 	}
 	/**
 	 * @param latitude the latitude to set
 	 */
-	public void setLatitude(BigDecimal latitude) {
+	public void setLatitude(Integer latitude) {
 		this.latitude = latitude;
 	}
 	/**
 	 * @return the longitude
 	 */
-	public BigDecimal getLongitude() {
+	public Integer getLongitude() {
 		return longitude;
 	}
 	/**
 	 * @param longitude the longitude to set
 	 */
-	public void setLongitude(BigDecimal longitude) {
+	public void setLongitude(Integer longitude) {
 		this.longitude = longitude;
 	}
 	/**
